Test for the create_collection route function
