+++ internal/api/store/collections_test.go
@@ -2,8 +2,6 @@ package store_test
 
 import (
 	"context"
-	"github.com/google/uuid"
-	"github.com/jackc/pgx/v5"
 	"github.com/pennsieve/collections-service/internal/api/store"
 	"github.com/pennsieve/collections-service/internal/dbmigrate"
 	"github.com/pennsieve/collections-service/internal/shared/logging"
@@ -29,7 +27,7 @@ func TestStore(t *testing.T) {
 	require.NoError(t, migrator.Up())
 	dbmigratetest.Close(t, migrator)
 
-	for scenario, tstFunc := range map[string]func(t *testing.T, collectionsStore *store.RDSCollectionsStore, expectationDB *ExpectationDB){
+	for scenario, tstFunc := range map[string]func(t *testing.T, collectionsStore *store.RDSCollectionsStore, expectationDB *fixtures.ExpectationDB){
 		"create collection, nil DOIs":          createCollectionNilDOIs,
 		"create collection, empty DOIs":        createCollectionEmptyDOIs,
 		"create collection, one DOI":           createCollectionOneDOI,
@@ -46,15 +44,15 @@ func TestStore(t *testing.T) {
 
 			collectionsStore := store.NewRDSCollectionsStore(db, config.CollectionsDatabase, logging.Default)
 
-			tstFunc(t, collectionsStore, NewExpectationDB(db, config.CollectionsDatabase))
+			tstFunc(t, collectionsStore, fixtures.NewExpectationDB(db, config.CollectionsDatabase))
 		})
 	}
 }
 
-func createCollectionNilDOIs(t *testing.T, store *store.RDSCollectionsStore, expectationDB *ExpectationDB) {
+func createCollectionNilDOIs(t *testing.T, store *store.RDSCollectionsStore, expectationDB *fixtures.ExpectationDB) {
 	ctx := context.Background()
 	expectedOwnerID := test.User.ID
-	expectedCollection := NewExpectedCollection().WithUser(expectedOwnerID, pgdb.Owner)
+	expectedCollection := fixtures.NewExpectedCollection().WithUser(expectedOwnerID, pgdb.Owner)
 
 	resp, err := store.CreateCollection(ctx, expectedOwnerID, expectedCollection.NodeID, expectedCollection.Name, expectedCollection.Description, nil)
 	require.NoError(t, err)
@@ -64,10 +62,10 @@ func createCollectionNilDOIs(t *testing.T, store *store.RDSCollectionsStore, exp
 	expectationDB.RequireCollection(ctx, t, expectedCollection, resp.ID)
 }
 
-func createCollectionEmptyDOIs(t *testing.T, store *store.RDSCollectionsStore, expectationDB *ExpectationDB) {
+func createCollectionEmptyDOIs(t *testing.T, store *store.RDSCollectionsStore, expectationDB *fixtures.ExpectationDB) {
 	ctx := context.Background()
 	expectedOwnerID := test.User.ID
-	expectedCollection := NewExpectedCollection().WithUser(expectedOwnerID, pgdb.Owner)
+	expectedCollection := fixtures.NewExpectedCollection().WithUser(expectedOwnerID, pgdb.Owner)
 
 	resp, err := store.CreateCollection(ctx, expectedOwnerID, expectedCollection.NodeID, expectedCollection.Name, expectedCollection.Description, []string{})
 	require.NoError(t, err)
@@ -77,10 +75,10 @@ func createCollectionEmptyDOIs(t *testing.T, store *store.RDSCollectionsStore, e
 	expectationDB.RequireCollection(ctx, t, expectedCollection, resp.ID)
 }
 
-func createCollectionOneDOI(t *testing.T, store *store.RDSCollectionsStore, expectationDB *ExpectationDB) {
+func createCollectionOneDOI(t *testing.T, store *store.RDSCollectionsStore, expectationDB *fixtures.ExpectationDB) {
 	ctx := context.Background()
 	expectedOwnerID := test.User.ID
-	expectedCollection := NewExpectedCollection().WithUser(expectedOwnerID, pgdb.Owner).WithDOIs(test.NewPennsieveDOI())
+	expectedCollection := fixtures.NewExpectedCollection().WithUser(expectedOwnerID, pgdb.Owner).WithDOIs(test.NewPennsieveDOI())
 
 	resp, err := store.CreateCollection(ctx, expectedOwnerID, expectedCollection.NodeID, expectedCollection.Name, expectedCollection.Description, expectedCollection.DOIs.Strings())
 	require.NoError(t, err)
@@ -91,10 +89,10 @@ func createCollectionOneDOI(t *testing.T, store *store.RDSCollectionsStore, expe
 
 }
 
-func createCollectionManyDOIs(t *testing.T, store *store.RDSCollectionsStore, expectationDB *ExpectationDB) {
+func createCollectionManyDOIs(t *testing.T, store *store.RDSCollectionsStore, expectationDB *fixtures.ExpectationDB) {
 	ctx := context.Background()
 	expectedOwnerID := test.User.ID
-	expectedCollection := NewExpectedCollection().WithUser(expectedOwnerID, pgdb.Owner).WithDOIs(test.NewPennsieveDOI(), test.NewPennsieveDOI(), test.NewPennsieveDOI())
+	expectedCollection := fixtures.NewExpectedCollection().WithUser(expectedOwnerID, pgdb.Owner).WithDOIs(test.NewPennsieveDOI(), test.NewPennsieveDOI(), test.NewPennsieveDOI())
 
 	resp, err := store.CreateCollection(ctx, expectedOwnerID, expectedCollection.NodeID, expectedCollection.Name, expectedCollection.Description, expectedCollection.DOIs.Strings())
 	require.NoError(t, err)
@@ -105,10 +103,10 @@ func createCollectionManyDOIs(t *testing.T, store *store.RDSCollectionsStore, ex
 
 }
 
-func createCollectionEmptyDescription(t *testing.T, store *store.RDSCollectionsStore, expectationDB *ExpectationDB) {
+func createCollectionEmptyDescription(t *testing.T, store *store.RDSCollectionsStore, expectationDB *fixtures.ExpectationDB) {
 	ctx := context.Background()
 	expectedOwnerID := test.User.ID
-	expectedCollection := NewExpectedCollection().WithUser(expectedOwnerID, pgdb.Owner).WithDOIs(test.NewPennsieveDOI(), test.NewPennsieveDOI(), test.NewPennsieveDOI())
+	expectedCollection := fixtures.NewExpectedCollection().WithUser(expectedOwnerID, pgdb.Owner).WithDOIs(test.NewPennsieveDOI(), test.NewPennsieveDOI(), test.NewPennsieveDOI())
 	expectedCollection.Description = ""
 
 	resp, err := store.CreateCollection(ctx, expectedOwnerID, expectedCollection.NodeID, expectedCollection.Name, expectedCollection.Description, expectedCollection.DOIs.Strings())
@@ -118,107 +116,4 @@ func createCollectionEmptyDescription(t *testing.T, store *store.RDSCollectionsS
 
 	expectationDB.RequireCollection(ctx, t, expectedCollection, resp.ID)
 
-}
-
-type ExpectedCollection struct {
-	Name        string
-	Description string
-	NodeID      string
-	Users       []ExpectedUser
-	DOIs        ExpectedDOIs
-}
-
-func NewExpectedCollection() *ExpectedCollection {
-	return &ExpectedCollection{
-		Name:        uuid.NewString(),
-		Description: uuid.NewString(),
-		NodeID:      uuid.NewString(),
-	}
-}
-
-type ExpectedUser struct {
-	UserID        int64
-	PermissionBit pgdb.DbPermission
-}
-
-func (c *ExpectedCollection) WithUser(userID int64, permission pgdb.DbPermission) *ExpectedCollection {
-	c.Users = append(c.Users, ExpectedUser{userID, permission})
-	return c
-}
-
-type ExpectedDOI struct {
-	DOI string
-}
-
-func (c *ExpectedCollection) WithDOIs(dois ...string) *ExpectedCollection {
-	for _, doi := range dois {
-		c.DOIs = append(c.DOIs, ExpectedDOI{DOI: doi})
-	}
-	return c
-}
-
-type ExpectedDOIs []ExpectedDOI
-
-func (d ExpectedDOIs) Strings() []string {
-	if len(d) == 0 {
-		return nil
-	}
-	strs := make([]string, len(d))
-	for i, doi := range d {
-		strs[i] = doi.DOI
-	}
-	return strs
-}
-
-type ExpectationDB struct {
-	db     *test.PostgresDB
-	dbName string
-}
-
-func NewExpectationDB(db *test.PostgresDB, dbName string) *ExpectationDB {
-	return &ExpectationDB{
-		db:     db,
-		dbName: dbName,
-	}
-}
-
-func (e *ExpectationDB) Connect(ctx context.Context, t require.TestingT) *pgx.Conn {
-	test.Helper(t)
-	conn, err := e.db.Connect(ctx, e.dbName)
-	require.NoError(t, err)
-	return conn
-}
-
-func (e *ExpectationDB) RequireCollection(ctx context.Context, t require.TestingT, expected *ExpectedCollection, expectedCollectionID int64) {
-	test.Helper(t)
-	conn := e.Connect(ctx, t)
-	defer test.CloseConnection(ctx, t, conn)
-
-	actual := fixtures.GetCollection(ctx, t, conn, expectedCollectionID)
-	require.Equal(t, expected.Name, actual.Name)
-	require.Equal(t, expected.Description, actual.Description)
-	require.Equal(t, expected.NodeID, actual.NodeID)
-	require.NotZero(t, actual.CreatedAt)
-	require.NotZero(t, actual.UpdatedAt)
-
-	actualUsers := fixtures.GetCollectionUsers(ctx, t, conn, expectedCollectionID)
-	require.Len(t, actualUsers, len(expected.Users))
-	for _, expectedUser := range expected.Users {
-		require.Contains(t, actualUsers, expectedUser.UserID)
-		actualUser := actualUsers[expectedUser.UserID]
-		require.Equal(t, expectedUser.PermissionBit, actualUser.PermissionBit)
-		require.Equal(t, expectedUser.PermissionBit.ToRole(), actualUser.Role.AsRole())
-		require.NotZero(t, actualUser.CreatedAt)
-		require.NotZero(t, actualUser.UpdatedAt)
-	}
-
-	actualDOIs := fixtures.GetDOIs(ctx, t, conn, expectedCollectionID)
-	require.Len(t, actualDOIs, len(expected.DOIs))
-	for _, expectedDOI := range expected.DOIs {
-		require.Contains(t, actualDOIs, expectedDOI.DOI)
-		actualDOI := actualDOIs[expectedDOI.DOI]
-		require.Equal(t, expectedDOI.DOI, actualDOI.DOI)
-		require.NotZero(t, actualDOI.CreatedAt)
-		require.NotZero(t, actualDOI.UpdatedAt)
-	}
 }
