+++ internal/test/fixtures/expectationdb.go
@@ -1,0 +1,113 @@
+package fixtures
+
+import (
+	"context"
+	"github.com/google/uuid"
+	"github.com/jackc/pgx/v5"
+	"github.com/pennsieve/collections-service/internal/test"
+	"github.com/pennsieve/pennsieve-go-core/pkg/models/pgdb"
+	"github.com/stretchr/testify/require"
+)
+
+type ExpectedCollection struct {
+	Name        string
+	Description string
+	NodeID      string
+	Users       []ExpectedUser
+	DOIs        ExpectedDOIs
+}
+
+func NewExpectedCollection() *ExpectedCollection {
+	return &ExpectedCollection{
+		Name:        uuid.NewString(),
+		Description: uuid.NewString(),
+		NodeID:      uuid.NewString(),
+	}
+}
+
+type ExpectedUser struct {
+	UserID        int64
+	PermissionBit pgdb.DbPermission
+}
+
+func (c *ExpectedCollection) WithUser(userID int64, permission pgdb.DbPermission) *ExpectedCollection {
+	c.Users = append(c.Users, ExpectedUser{userID, permission})
+	return c
+}
+
+type ExpectedDOI struct {
+	DOI string
+}
+
+func (c *ExpectedCollection) WithDOIs(dois ...string) *ExpectedCollection {
+	for _, doi := range dois {
+		c.DOIs = append(c.DOIs, ExpectedDOI{DOI: doi})
+	}
+	return c
+}
+
+type ExpectedDOIs []ExpectedDOI
+
+func (d ExpectedDOIs) Strings() []string {
+	if len(d) == 0 {
+		return nil
+	}
+	strs := make([]string, len(d))
+	for i, doi := range d {
+		strs[i] = doi.DOI
+	}
+	return strs
+}
+
+type ExpectationDB struct {
+	db     *test.PostgresDB
+	dbName string
+}
+
+func NewExpectationDB(db *test.PostgresDB, dbName string) *ExpectationDB {
+	return &ExpectationDB{
+		db:     db,
+		dbName: dbName,
+	}
+}
+
+func (e *ExpectationDB) Connect(ctx context.Context, t require.TestingT) *pgx.Conn {
+	test.Helper(t)
+	conn, err := e.db.Connect(ctx, e.dbName)
+	require.NoError(t, err)
+	return conn
+}
+
+func (e *ExpectationDB) RequireCollection(ctx context.Context, t require.TestingT, expected *ExpectedCollection, expectedCollectionID int64) {
+	test.Helper(t)
+	conn := e.Connect(ctx, t)
+	defer test.CloseConnection(ctx, t, conn)
+
+	actual := GetCollection(ctx, t, conn, expectedCollectionID)
+	require.Equal(t, expected.Name, actual.Name)
+	require.Equal(t, expected.Description, actual.Description)
+	require.Equal(t, expected.NodeID, actual.NodeID)
+	require.NotZero(t, actual.CreatedAt)
+	require.NotZero(t, actual.UpdatedAt)
+
+	actualUsers := GetCollectionUsers(ctx, t, conn, expectedCollectionID)
+	require.Len(t, actualUsers, len(expected.Users))
+	for _, expectedUser := range expected.Users {
+		require.Contains(t, actualUsers, expectedUser.UserID)
+		actualUser := actualUsers[expectedUser.UserID]
+		require.Equal(t, expectedUser.PermissionBit, actualUser.PermissionBit)
+		require.Equal(t, expectedUser.PermissionBit.ToRole(), actualUser.Role.AsRole())
+		require.NotZero(t, actualUser.CreatedAt)
+		require.NotZero(t, actualUser.UpdatedAt)
+	}
+
+	actualDOIs := GetDOIs(ctx, t, conn, expectedCollectionID)
+	require.Len(t, actualDOIs, len(expected.DOIs))
+	for _, expectedDOI := range expected.DOIs {
+		require.Contains(t, actualDOIs, expectedDOI.DOI)
+		actualDOI := actualDOIs[expectedDOI.DOI]
+		require.Equal(t, expectedDOI.DOI, actualDOI.DOI)
+		require.NotZero(t, actualDOI.CreatedAt)
+		require.NotZero(t, actualDOI.UpdatedAt)
+	}
+}
